% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(folder_path = file.path(getwd(), "data"))
}
\arguments{
\item{folder_path}{The path to the folder where SnackApp usage data is located. This path defaults to a folder named "data" within the current work directory.}
}
\value{
A list of tibbles which can then be analysed using other SnackApp functions.
}
\description{
load_data
}
