% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_summary.R
\name{participant_summary}
\alias{participant_summary}
\title{participant_summary}
\usage{
participant_summary(folder_path, csv = FALSE, r_object = TRUE)
}
\arguments{
\item{r_object}{}
}
\value{

}
\description{
participant_summary
}
