% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackapp_individual.R
\name{snack_app_individual}
\alias{snack_app_individual}
\title{Title}
\usage{
snack_app_individual(
  folder_path,
  csv = FALSE,
  r_object = TRUE,
  output_path = getwd()
)
}
\arguments{
\item{output_path}{}
}
\value{

}
\description{
Title
}
