% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackapp_day_summary.R
\name{snackapp_day_summary}
\alias{snackapp_day_summary}
\title{snackapp_day_summary}
\usage{
snackapp_day_summary(
  folder_path,
  csv = FALSE,
  r_object = TRUE,
  output_path = getwd()
)
}
\arguments{
\item{output_path}{}
}
\value{

}
\description{
snackapp_day_summary
}
