% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_summary.R
\name{day_summary}
\alias{day_summary}
\title{day_summary}
\usage{
day_summary(
  folder_path = file.path(getwd(), "data"),
  csv = FALSE,
  output_path = file.path(getwd(), "summary")
)
}
\arguments{
\item{folder_path}{A path to the folder which contains the SnackApp usage data. The file should only contain csv files.
Defaults to a folder in the current work directory named "data".}

\item{csv}{A logical variable that decides whether a csv output file is created.}

\item{output_path}{A path to a folder where the csv summary file(s) will be written. Defaults to a folder in the current
work directory named "summary".}
}
\value{
A dataframe that summarises the SnackApp usage data for each day a participant used the device.
}
\description{
Summarising the usage data for a participant on each day they used the intervention. A single tibble is
generated.
}
\examples{

}
