% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_summary.R
\name{individual_summary}
\alias{individual_summary}
\title{individual_summary}
\usage{
individual_summary(
  data,
  csv = FALSE,
  output_path = file.path(getwd(), "summary")
)
}
\arguments{
\item{data}{The SnackApp usage data that is to be analysed. This parameter accepts the output of the load_data function.
It also accepts a dataframe with 6 columns: id, Date, Event, Metric, Goal Is Automatic Set and Goal Value Set.}

\item{csv}{A logical variable that decides whether a csv output file is created.}

\item{output_path}{A path to a folder where the csv summary file(s) will be written. Defaults to a folder in the current
work directory named "summary".}
}
\value{
A dataframe is generated to summarise each participants usage data.
}
\description{
Creating a separate dataframe for each participant for which data has been collected. A list of tibbles are generated while
separate .csv files can be produced and written to file.
}
