% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_summary.R
\name{participant_summary}
\alias{participant_summary}
\title{participant_summary}
\usage{
participant_summary(
  data,
  csv = FALSE,
  output_path = file.path(getwd(), "summary")
)
}
\arguments{
\item{data}{The SnackApp usage data that is to be analysed. This parameter accepts the output of the load_data function.
It also accepts a dataframe with 6 columns: id, Date, Event, Metric, Goal Is Automatic Set and Goal Value Set.}

\item{csv}{A logical variable that decides whether a csv output file is created.}

\item{output_path}{A path to a folder where the csv summary file(s) will be written. Defaults to a folder in the current
work directory named "summary".}
}
\value{
A dataframe that summarises the SnackApp usage data. The summary is generated with one row per participant.
}
\description{
Summarises participants usage over all days they are part of the intervention. A tibble is generated with one row per
participant.
}
